requirements.txt
aiohttp==3.13.0
aiohttp-retry==2.9.1
base58==2.1.1
python-dotenv==1.1.0
requests==2.32.3
solders==0.26.0
